+++ src/protocol/machine_driver.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2018 Baidu, Inc. All Rights Reserved.
+// Copyright (c) 2018-present Baidu, Inc. All Rights Reserved.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -56,7 +56,7 @@ void* MachineDriver::bthread_callback(void* void_arg) {
     }
 
     task->socket->thread_idx = thread_idx;
-    if (thread_idx >= _driver->_thread_num) {
+    if ((uint32_t)thread_idx >= _driver->_thread_num) {
         DB_FATAL("tid: %d, thread_idx: %d, _thread_num: %d", tid, thread_idx, _driver->_thread_num);
         exit(-1);
     }
